+++ src/components/ui/MyPetition/index.tsx
@@ -1,6 +1,6 @@
 import PetitionList from 'components/common/PetitionList';
 import RadioTabMenu from 'components/common/RadioTabMenu';
-import { MyPetitionFeature } from 'features/MyPetitionFeature';
+import { useMyPetitionList } from 'features/MyPetitionFeature';
 import NotFound from 'pages/404';
 import Loading from 'pages/Loading';
 import { useState } from 'react';
@@ -9,13 +9,13 @@ import * as S from './style';
 
 const MyPetition = () => {
   const [status, setStatus] = useState<StatusType>('WROTE');
-  const { useMyPetitionList } = MyPetitionFeature(status);
 
-  const { data, isError, isLoading } = useMyPetitionList();
+  const { data, isError, isLoading } = useMyPetitionList(status);
 
   return (
     <S.MyPetitionLayout>
       <S.MyPetitionWrap>
+        ;
         <S.InnerWrap>
           <S.RadioTabMenuWrap>
             <RadioTabMenu
