+++ src/features/LoginFeature.ts
@@ -5,37 +5,18 @@ import { loginUser } from 'api/auth.api';
 import { LoginType } from 'types/auth.type';
 import { Storage } from 'lib/storage/storage';
 import { toast } from 'react-toastify';
-import { ChangeEvent, useCallback, useState } from 'react';
 
-export const LoginFeature = () => {
-  const [loginData, setLoginData] = useState<LoginType>({
-    username: '',
-    password: '',
+export const useLoginMutation = (loginData: LoginType) => {
+  const navigate = useNavigate();
+
+  return useMutation(() => loginUser(loginData), {
+    onSuccess: (res) => {
+      const { accessToken, refreshToken } = res;
+
+      Storage.setItem(ACCESS_KEY, accessToken);
+      Storage.setItem(REFRESH_KEY, refreshToken);
+      toast.success('로그인 성공');
+      navigate('/');
+    },
   });
-
-  const handleLoginData = (e: ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = e.target;
-    setLoginData({ ...loginData, [name]: value });
-  };
-
-  const useLoginMutation = useCallback(() => {
-    const navigate = useNavigate();
-
-    return useMutation(() => loginUser(loginData), {
-      onSuccess: (res) => {
-        const { accessToken, refreshToken } = res;
-
-        Storage.setItem(ACCESS_KEY, accessToken);
-        Storage.setItem(REFRESH_KEY, refreshToken);
-        toast.success('로그인 성공');
-        navigate('/');
-      },
-    });
-  }, []);
-
-  return {
-    handleLoginData,
-    useLoginMutation,
-    loginData,
-  };
 };
