+++ src/features/LogoutFeature.ts
@@ -4,18 +4,14 @@ import { useSetRecoilState } from 'recoil';
 import { userEmpty, userState } from 'atoms/user.atom';
 import { useCallback } from 'react';
 
-export const LogoutFeature = () => {
+export const useLogoutMutation = useCallback(() => {
   const setUser = useSetRecoilState(userState);
 
-  const useLogoutMutation = useCallback(() => {
-    return useMutation(logoutUser, {
-      onSuccess: () => {
-        localStorage.clear();
-        setUser(userEmpty);
-        window.location.href = '/';
-      },
-    });
-  }, []);
-
-  return { useLogoutMutation };
-};
+  return useMutation(logoutUser, {
+    onSuccess: () => {
+      localStorage.clear();
+      setUser(userEmpty);
+      window.location.href = '/';
+    },
+  });
+}, []);
